Traktor: Flash player stereoscopic support. Able to use "free" channel in DXT5nm compressed textures for additional alpha.
[git-p4: depot-paths = "//depot/traktor/": change = 6154]

+++ code/Flash/Editor/FlashPreviewControl.cpp
@@ -281,7 +281,7 @@ void FlashPreviewControl::eventPaint(ui::Event* event)
 		// Build render context.
 		if (m_movie)
 		{
-			m_displayRenderer->build(0);
+			m_displayRenderer->build(uint32_t(0));
 			m_moviePlayer->renderFrame();
 		}
